Aug 9 Commit 2 Cleanup and preparation for new feed class
