+++ nsound_old/tf_rnn.py
@@ -8,13 +8,26 @@ from tensorflow.contrib import rnn
 import numpy as np
 import os
 
-print('Loading MNIST data...')
+print('Loading Data...')
 
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 
 # Import MNIST data
 from tensorflow.examples.tutorials.mnist import input_data
 mnist = input_data.read_data_sets("/tmp/data/", one_hot=True)
+
+# PARAMS
+learning_rate = 0.001
+training_iters = 100000
+batch_size = 128
+display_step = 10
+#Data Shape
+n_input = 28
+n_steps = 28
+n_hidden = 128 # hidden layer num of features
+n_classtypes = 10 # MNIST total classes (0-9 digits)
+
+print(input_data)
 
 #Add sumaries to TensorBoard for weights and biases
 def var_summaries(var):
@@ -28,43 +41,24 @@ def var_summaries(var):
             tf.summary.scalar('min', tf.reduce_min(var))
             tf.summary.histogram('histogram', var)
 
-# Parameters
-learning_rate = 0.001
-training_iters = 100000
-batch_size = 128
-display_step = 10
-
-# Network Parameters
-n_input = 28 # MNIST data input (img shape: 28*28)
-n_steps = 28 # timesteps
-n_hidden = 128 # hidden layer num of features
-n_classes = 10 # MNIST total classes (0-9 digits)
-
 # tf Graph input
-x = tf.placeholder("float", [None, n_steps, n_input])
-y = tf.placeholder("float", [None, n_classes])
+input_vector = tf.placeholder("float", [None, n_steps, n_input])
+y = tf.placeholder("float", [None, n_classtypes])
 
-# Define weights
-# weights = {
-#     # Hidden layer weights => 2*n_hidden because of forward + backward cells
-#     'out': tf.Variable(tf.random_normal([2*n_hidden, n_classes]))
-# }
-# biases = {
-#     'out': tf.Variable(tf.random_normal([n_classes]))
-# }
-weights = tf.Variable(tf.random_normal([2*n_hidden, n_classes]))
-biases = tf.Variable(tf.random_normal([n_classes]))
+weights = tf.Variable(tf.random_normal([2*n_hidden, n_classtypes]))
+biases = tf.Variable(tf.random_normal([n_classtypes]))
 var_summaries(weights)
 var_summaries(biases)
 
-def BiRNN(x, weights, biases):
+def BiRNN(input_vector, weights, biases):
 
     # Prepare data shape to match `bidirectional_rnn` function requirements
     # Current data input shape: (batch_size, n_steps, n_input)
     # Required shape: 'n_steps' tensors list of shape (batch_size, n_input)
 
     # Unstack to get a list of 'n_steps' tensors of shape (batch_size, n_input)
-    x = tf.unstack(x, n_steps, 1)
+    input_vector = tf.unstack(input_vector, n_steps, 1)
+
 
     # Define lstm cells with tensorflow
     # Forward direction cell
@@ -73,18 +67,13 @@ def BiRNN(x, weights, biases):
     lstm_bw_cell = rnn.BasicLSTMCell(n_hidden, forget_bias=1.0)
 
     # Get lstm cell output
-    try:
-        outputs, _, _ = rnn.static_bidirectional_rnn(lstm_fw_cell, lstm_bw_cell, x,
-                                              dtype=tf.float32)
-    except Exception: # Old TensorFlow version only returns outputs not states
-        outputs = rnn.static_bidirectional_rnn(lstm_fw_cell, lstm_bw_cell, x,
-                                        dtype=tf.float32)
+    outputs,_,_ = rnn.static_bidirectional_rnn(lstm_fw_cell, lstm_bw_cell, input_vector, dtype=tf.float32)
 
     # Linear activation, using rnn inner loop last output
     #return tf.matmul(outputs[-1], weights['out']) + biases['out']
     return tf.matmul(outputs[-1], weights) + biases
 
-pred = BiRNN(x, weights, biases)
+pred = BiRNN(input_vector, weights, biases)
 
 # Define loss and optimizer
 with tf.name_scope('cross_entropy'):
@@ -110,26 +99,26 @@ with tf.Session() as sess:
     print("Starting Tensorboard...")
     writer = tf.summary.FileWriter("totalsummary", sess.graph)
 
-    # Keep training until reach max iterations
+    print("Training...")
     while step * batch_size < training_iters:
         batch_x, batch_y = mnist.train.next_batch(batch_size)
         # Reshape data to get 28 seq of 28 elements
         batch_x = batch_x.reshape((batch_size, n_steps, n_input))
         # Run optimization op (backprop)
-        sess.run(optimizer, feed_dict={x: batch_x, y: batch_y})
+        sess.run(optimizer, feed_dict={input_vector: batch_x, y: batch_y})
         if step % display_step == 0:
             run_options = tf.RunOptions(trace_level=tf.RunOptions.FULL_TRACE)
             run_metadata = tf.RunMetadata()
             # Calculate batch accuracy
-            acc = sess.run(accuracy, feed_dict={x: batch_x, y: batch_y})
+            acc = sess.run(accuracy, feed_dict={input_vector: batch_x, y: batch_y})
             i = tf.convert_to_tensor(step, dtype = tf.int8)
             # Calculate batch loss
-            loss = sess.run(cost, feed_dict={x: batch_x, y: batch_y})
+            loss = sess.run(cost, feed_dict={input_vector: batch_x, y: batch_y})
             print("Iter " + str(step*batch_size) + ", Minibatch Loss= " + \
                   "{:.6f}".format(loss) + ", Training Accuracy= " + \
                   "{:.5f}".format(acc))
 
-            trainsummary = sess.run(merged, feed_dict={x: batch_x, y: batch_y})
+            trainsummary = sess.run(merged, feed_dict={input_vector: batch_x, y: batch_y})
             writer.add_summary(trainsummary, step)
             writer.flush()
         step += 1
@@ -140,7 +129,7 @@ with tf.Session() as sess:
     test_data = mnist.test.images[:test_len].reshape((-1, n_steps, n_input))
     test_label = mnist.test.labels[:test_len]
     print("Testing Accuracy:", \
-        sess.run(accuracy, feed_dict={x: test_data, y: test_label}))
+        sess.run(accuracy, feed_dict={input_vector: test_data, y: test_label}))
