+++ tf_feed.py
@@ -1,0 +1,16 @@
+import numpy as np
+
+def dataset():
+    """
+    Summary:
+        #Class
+    Parameters:
+    Output:
+    """
+    def __init(self,sound_path,label):
+        return 0
+
+def data_loader():
+    return 0
+def feed_model():
+    return 0
